requirements.txt
scipy==1.17.1
sentence-transformers==5.4.1
langchain-chroma==1.1.0
python-dotenv==1.2.2
langchain-core==1.2.30
langchain-google-genai==4.2.2
